Changes for PSR-4 compliance
