+++ composer.json
@@ -28,6 +28,9 @@
         ]
     },
     "autoload" : {
-         "classmap": [ "src/"]
+        "psr-4": {
+            "Shihjay2\\" : "src"
+        },
+        "files": [ "src/functions.php" ]
     }
 }
