+++ src/OpenIDConnectUMAClient.php
@@ -31,64 +31,6 @@ use phpseclib\Crypt\RSA;
 
 /**
  *
- * JWT signature verification support by Jonathan Reed <jdreed@mit.edu>
- * Licensed under the same license as the rest of this file.
- *
- * phpseclib is required to validate the signatures of some tokens.
- * It can be downloaded from: http://phpseclib.sourceforge.net/
- */
-// include('Crypt/RSA.php');
-// use Crypt_RSA;
-
-if (!class_exists('phpseclib\Crypt\RSA')) {
-	user_error('Unable to find phpseclib Crypt/RSA.php.  Ensure phpseclib is installed and in include_path');
-}
-
-/**
- * A wrapper around base64_decode which decodes Base64URL-encoded data,
- * which is not the same alphabet as base64.
- */
-function base64url_decode($base64url) {
-	return base64_decode(b64url2b64($base64url));
-}
-
-/**
- * Per RFC4648, "base64 encoding with URL-safe and filename-safe
- * alphabet".  This just replaces characters 62 and 63.  None of the
- * reference implementations seem to restore the padding if necessary,
- * but we'll do it anyway.
- *
- */
-function b64url2b64($base64url) {
-	// "Shouldn't" be necessary, but why not
-	$padding = strlen($base64url) % 4;
-	if ($padding > 0) {
-	$base64url .= str_repeat("=", 4 - $padding);
-	}
-	return strtr($base64url, '-_', '+/');
-}
-
-
-/**
- * OpenIDConnect Exception Class
- */
-class OpenIDConnectClientException extends \Exception
-{
-
-}
-
-/**
- * Require the CURL and JSON PHP extentions to be installed
- */
-if (!function_exists('curl_init')) {
-	throw new OpenIDConnectClientException('OpenIDConnect needs the CURL PHP extension.');
-}
-if (!function_exists('json_decode')) {
-	throw new OpenIDConnectClientException('OpenIDConnect needs the JSON PHP extension.');
-}
-
-/**
- *
  * Please note this class stores nonces by default in $_SESSION['openid_connect_nonce']
  *
  */
