+++ src/functions.php
@@ -1,0 +1,48 @@
+<?php
+
+namespace Shihjay2;
+
+/**
+ *
+ * Copyright MITRE 2018
+ *
+ * OpenIDConnectClient for PHP5
+ * Author: Michael Jett <mjett@mitre.org>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License. You may obtain
+ * a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ *
+ */
+
+/**
+ * A wrapper around base64_decode which decodes Base64URL-encoded data,
+ * which is not the same alphabet as base64.
+ */
+function base64url_decode($base64url) {
+	return base64_decode(b64url2b64($base64url));
+}
+
+/**
+ * Per RFC4648, "base64 encoding with URL-safe and filename-safe
+ * alphabet".  This just replaces characters 62 and 63.  None of the
+ * reference implementations seem to restore the padding if necessary,
+ * but we'll do it anyway.
+ *
+ */
+function b64url2b64($base64url) {
+	// "Shouldn't" be necessary, but why not
+	$padding = strlen($base64url) % 4;
+	if ($padding > 0) {
+	$base64url .= str_repeat("=", 4 - $padding);
+	}
+	return strtr($base64url, '-_', '+/');
+}
